pyproject.toml
[project]
name = "searxng-mcp"
version = "0.1.0"
description = "FastMCP server for web search via SearXNG and page scraping"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"}
]
keywords = ["mcp", "fastmcp", "searxng", "web-scraping", "search"]
dependencies = [
    "fastmcp",
    "requests",
    "beautifulsoup4",
    "lxml",
    "playwright",
    "python-toon",
    "python-dotenv",
]

[project.scripts]
searxng-mcp = "server:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["searxng_mcp"]